#28 [RENAME] 더미데이터들 constants 폴더로 이동
